+++ rpcs3/Emu/Cell/SPUInterpreter.h
@@ -257,7 +257,7 @@ private:
 			return;
 		}
 
-		u64 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
+		u32 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
 		if (CPU.GPR[rt]._u32[3] == 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -276,7 +276,7 @@ private:
 			return;
 		}
 
-		u64 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
+		u32 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
 		if (CPU.GPR[rt]._u32[3] != 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -295,7 +295,7 @@ private:
 			return;
 		}
 
-		u64 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
+		u32 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
 		if (CPU.GPR[rt]._u16[6] == 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -314,7 +314,7 @@ private:
 			return;
 		}
 
-		u64 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
+		u32 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
 		if (CPU.GPR[rt]._u16[6] != 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -343,7 +343,7 @@ private:
 			return;
 		}
 
-		u64 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
+		u32 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
 		LOG5_OPCODE("branch (0x%llx)", target);
 		CPU.SetBranch(target);
 	}
@@ -355,9 +355,9 @@ private:
 			return;
 		}
 
-		u64 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
+		u32 target = branchTarget(CPU.GPR[ra]._u32[3], 0);
 		CPU.GPR[rt].clear();
-		CPU.GPR[rt]._u32[3] = (u32)CPU.PC + 4;		
+		CPU.GPR[rt]._u32[3] = CPU.PC + 4;		
 		LOG5_OPCODE("branch (0x%llx)", target);
 		CPU.SetBranch(target);
 	}
@@ -1100,7 +1100,7 @@ private:
 	//0 - 8
 	void BRZ(u32 rt, s32 i16)
 	{
-		u64 target = branchTarget(CPU.PC, i16);
+		u32 target = branchTarget(CPU.PC, i16);
 		if (CPU.GPR[rt]._u32[3] == 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -1119,7 +1119,7 @@ private:
 	}
 	void BRNZ(u32 rt, s32 i16)
 	{
-		u64 target = branchTarget(CPU.PC, i16);
+		u32 target = branchTarget(CPU.PC, i16);
 		if (CPU.GPR[rt]._u32[3] != 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -1132,7 +1132,7 @@ private:
 	}
 	void BRHZ(u32 rt, s32 i16)
 	{
-		u64 target = branchTarget(CPU.PC, i16);
+		u32 target = branchTarget(CPU.PC, i16);
 		if (CPU.GPR[rt]._u16[6] == 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -1145,7 +1145,7 @@ private:
 	}
 	void BRHNZ(u32 rt, s32 i16)
 	{
-		u64 target = branchTarget(CPU.PC, i16);
+		u32 target = branchTarget(CPU.PC, i16);
 		if (CPU.GPR[rt]._u16[6] != 0)
 		{
 			LOG5_OPCODE("taken (0x%llx)", target);
@@ -1164,7 +1164,7 @@ private:
 	}
 	void BRA(s32 i16)
 	{
-		u64 target = branchTarget(0, i16);
+		u32 target = branchTarget(0, i16);
 		LOG5_OPCODE("branch (0x%llx)", target);
 		CPU.SetBranch(target);
 	}
@@ -1176,15 +1176,15 @@ private:
 	}
 	void BRASL(u32 rt, s32 i16)
 	{
-		u64 target = branchTarget(0, i16);
+		u32 target = branchTarget(0, i16);
 		CPU.GPR[rt].clear();
-		CPU.GPR[rt]._u32[3] = (u32)CPU.PC + 4;
+		CPU.GPR[rt]._u32[3] = CPU.PC + 4;
 		LOG5_OPCODE("branch (0x%llx)", target);
 		CPU.SetBranch(target);
 	}
 	void BR(s32 i16)
 	{
-		u64 target = branchTarget(CPU.PC, i16);
+		u32 target = branchTarget(CPU.PC, i16);
 		LOG5_OPCODE("branch (0x%llx)", target);
 		CPU.SetBranch(target);
 	}
@@ -1206,9 +1206,9 @@ private:
 	}
 	void BRSL(u32 rt, s32 i16)
 	{
-		u64 target = branchTarget(CPU.PC, i16);
+		u32 target = branchTarget(CPU.PC, i16);
 		CPU.GPR[rt].clear();
-		CPU.GPR[rt]._u32[3] = (u32)CPU.PC + 4;
+		CPU.GPR[rt]._u32[3] = CPU.PC + 4;
 		LOG5_OPCODE("branch (0x%llx)", target);
 		CPU.SetBranch(target);
 	}
